+++ lib/ripe.rb
@@ -1,8 +1,6 @@
 require_relative 'ripe/blocks'
-require_relative 'ripe/db'
 require_relative 'ripe/dsl'
 require_relative 'ripe/library'
-require_relative 'ripe/repo'
 require_relative 'ripe/worker_controller'
 
 require_relative 'ripe/cli'
@@ -10,4 +8,5 @@ require_relative 'ripe/version'
 
 module Ripe
   PATH = File.expand_path('..', File.dirname(__FILE__))
+  REPOSITORY_PATH = '.ripe'
 end