+++ lib/ripe/db.rb
@@ -1,7 +1,0 @@
-require 'active_record'
-require 'fileutils'
-
-require_relative 'db/task'
-require_relative 'db/task_migration'
-require_relative 'db/worker'
-require_relative 'db/worker_migration'