+++ spec/spec_helper.rb
@@ -5,5 +5,4 @@ end
 
 require_relative '../lib/ripe'
 require_relative 'testpack'
-
-include Ripe
+require 'tempfile'