+++ spec/worker_controller_spec.rb
@@ -1,9 +1,10 @@
 require 'spec_helper'
 
-include Ripe::DSL # required by dirty hack +WorkerController#prepare+
+include Ripe::DSL
 
 require 'digest'
 require 'fileutils'
+require 'tempfile'
 
 def signature(file)
   Digest::MD5.hexdigest(File.read(file).gsub(/LOG=.*\n/, '').gsub(/exec 1>".*\n/, '') )
@@ -19,11 +20,7 @@ describe WorkerController do
       Dir.chdir(@tmpdir)
 
       ENV['RIPELIB'] = @test.lib_path
-      @repo = Repo.new
-      @repo.attach_or_create
       @library = Library
-      @controller = @repo.controller
-
       @test.samples.each do |sample|
         FileUtils.mkdir_p(sample)
         @test.steps.each do |step|
@@ -32,6 +29,7 @@ describe WorkerController do
           FileUtils.cp(source, dest)
         end
       end
+      Dir.mkdir REPOSITORY_PATH
     end
 
     after :all do
@@ -41,86 +39,90 @@ describe WorkerController do
 
     describe '#prepare' do
       it 'prepares workers' do
-        workers = @controller.prepare 'foobar', @test.samples, pwd: @test.path, mode: :force
+        workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
         # Prepares workers 1-2-3
-        expect(DB::Worker.all.length).to eql 3
-        # Returns the workers from the call to prepare
-        expect(workers).to eql DB::Worker.all.to_a
+        expect(workers.length).to eql 3
       end
 
       it 'prepares workers with accurate task scripts' do
-        DB::Task.all.each do |task|
-          test_hash = signature(task.sh)
-          ref_hash = signature("#{@test.path}/#{task.sh}")
+        workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
+
+        workers.map { |a| a.tasks }.inject(&:+).each do |task|
+          test_hash = signature(task.log)
+          ref_hash = signature("#{@test.path}/#{task.log}")
           expect(test_hash).to eql ref_hash
         end
       end
 
-      it 'properly prepares workers in force mode' do
-        sample = @test.samples[0]
+      # it 'properly prepares workers in force mode' do
+      #   workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
+      #   workers += WorkerController.new('foobar', [@test.samples[0]], '.ripe/workers', pwd: @test.path, mode: :force).workers
 
-        @controller.prepare 'foobar', [sample], pwd: @test.path, mode: :force
+      #   ref_tasks = workers[0].tasks
+      #   test_tasks = workers[3].tasks
 
-        ref_tasks = DB::Worker.find(1).tasks
-        test_tasks = DB::Worker.find(4).tasks
+      #   expect(ref_tasks.length).to eql 3
+      #   expect(test_tasks.length).to eql 3
 
-        expect(ref_tasks.length).to eql 3
-        expect(test_tasks.length).to eql 3
+      #   ref_tasks.zip(test_tasks).map do |ref, test|
+      #     ref_hash = signature(ref.log)
+      #     test_hash = signature(test.log)
+      #     expect(test_hash).to eql ref_hash
+      #   end
+      # end
 
-        ref_tasks.zip(test_tasks).map do |ref, test|
-          ref_hash = signature(ref.sh)
-          test_hash = signature(test.sh)
-          expect(test_hash).to eql ref_hash
-        end
-      end
+      # it 'properly prepares workers in patch mode' do
+      #   sample = @test.samples[0]
+      #   step = @test.steps[1]
 
-      it 'properly prepares workers in patch mode' do
-        sample = @test.samples[0]
-        step = @test.steps[1]
+      #   workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
 
-        # Delete the first output
-        FileUtils.rm_r("#{sample}/#{step}")
+      #   # Delete the first output
+      #   FileUtils.rm_r("#{sample}/#{step}")
 
-        @controller.prepare 'foobar', [sample], pwd: @test.path, mode: :patch
+      #   workers += WorkerController.new('foobar', [sample], '.ripe/workers', pwd: @test.path, mode: :patch).workers
 
-        ref_tasks = DB::Worker.find(1).tasks
-        test_tasks = DB::Worker.find(5).tasks
+      #   ref_tasks = workers[0].tasks
+      #   test_tasks = workers[3].tasks
 
-        expect(ref_tasks.length).to eql 3
-        expect(test_tasks.length).to eql 1
+      #   expect(ref_tasks.length).to eql 3
+      #   expect(test_tasks.length).to eql 1
 
-        ref_hash = signature(ref_tasks.first.sh)
-        test_hash = signature(test_tasks.first.sh)
+      #   ref_hash = signature(ref_tasks.first.log)
+      #   test_hash = signature(test_tasks.first.log)
 
-        expect(test_hash).to eql ref_hash
-      end
+      #   expect(test_hash).to eql ref_hash
+      # end
 
-      it 'properly prepares workers in depend mode' do
-        sample = @test.samples[1]
-        step = @test.steps[1]
+      # it 'properly prepares workers in depend mode' do
+      #   sample = @test.samples[1]
+      #   step = @test.steps[1]
 
-        # Delete the first output
-        FileUtils.rm_r("#{sample}/#{step}")
+      #   workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
 
-        @controller.prepare 'foobar', [sample], pwd: @test.path, mode: :depend
+      #   # Delete the first output
+      #   FileUtils.rm_r("#{sample}/#{step}")
 
-        ref_tasks = DB::Worker.find(2).tasks
-        test_tasks = DB::Worker.find(6).tasks
+      #   workers += WorkerController.new('foobar', [sample], '.ripe/workers', pwd: @test.path, mode: :depend).workers
 
-        expect(ref_tasks.length).to eql 3
-        expect(test_tasks.length).to eql 3
+      #   ref_tasks = workers[1].tasks
+      #   test_tasks = workers[3].tasks
 
-        ref_tasks.zip(test_tasks).map do |ref, test|
-          ref_hash = signature(ref.sh)
-          test_hash = signature(test.sh)
-          expect(test_hash).to eql ref_hash
-        end
-      end
+      #   expect(ref_tasks.length).to eql 3
+      #   expect(test_tasks.length).to eql 3
+
+      #   ref_tasks.zip(test_tasks).map do |ref, test|
+      #     ref_hash = signature(ref.log)
+      #     test_hash = signature(test.log)
+      #     expect(test_hash).to eql ref_hash
+      #   end
+      # end
 
       describe '#local' do
         it 'runs worker jobs locally' do
-          worker = DB::Worker.find(1)
-          @controller.local worker
+          workers = WorkerController.new('foobar', @test.samples, '.ripe/workers', pwd: @test.path, mode: :force).workers
+          worker = workers[0]
+          `bash #{worker.sh}`
           @test.steps.map do |step|
             test_hash = signature("#{@test.samples[0]}/#{step}")
             ref_hash = signature("#{@test.path}/#{@test.samples[0]}/#{step}")
@@ -129,15 +131,6 @@ describe WorkerController do
         end
       end
 
-      describe '#sync' do
-        it 'does not affected unprepared workers' do
-          before_statuses = DB::Worker.where(status: :prepared).map(&:id)
-          @controller.sync
-          after_statuses = DB::Worker.where(status: :prepared).map(&:id)
-
-          expect(after_statuses).to eql before_statuses
-        end
-      end
     end
   end
 end